# internal/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BevShop API",
    "version": "1.0.0",
    "description": "Single-owner beverage shop API. Use auth/setup once, or auth/login, then copy the returned token into Authorize (without the Bearer prefix). Request IDs must reference existing records. Dates use RFC3339 timestamps with an explicit timezone. This document describes the currently implemented routes."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List products",
        "operationId": "product_list",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/product_Product"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create product",
        "operationId": "product_create",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/product_Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Product already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/product_CreateRequest"
              },
              "example": {
                "name": "Coca-Cola 500ml",
                "category_id": "11111111-1111-4111-8111-111111111111",
                "description": "",
                "units_per_pack": 24,
                "selling_price": 700,
                "low_stock_level": 20
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product",
        "operationId": "product_get",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/product_Product"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product",
        "operationId": "product_update",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/product_Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Product already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/product_UpdateRequest"
              },
              "example": {
                "name": "Coca-Cola 500ml",
                "category_id": "11111111-1111-4111-8111-111111111111",
                "description": "",
                "units_per_pack": 24,
                "selling_price": 700,
                "low_stock_level": 20
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/deactivate": {
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "Deactivate product",
        "operationId": "product_deactivate",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
...
    "/api/v1/purchases": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "summary": "List purchases",
        "operationId": "purchase_list",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/purchase_Purchase"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Inclusive start, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-10-01T00:00:00+02:00"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Exclusive end, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-11-01T00:00:00+02:00"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum records to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of records to skip.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "supplier_id",
            "in": "query",
            "required": false,
            "description": "Optional filter.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Purchases"
        ],
        "summary": "Create purchase",
        "operationId": "purchase_create",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/purchase_Purchase"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/purchase_CreateRequest"
              },
              "example": {
                "supplier_id": "11111111-1111-4111-8111-111111111111",
                "notes": "",
                "items": [
                  {
                    "product_id": "11111111-1111-4111-8111-111111111111",
                    "packs": 10,
                    "price_per_pack": 12000
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/purchases/{id}": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "summary": "Get purchase",
        "operationId": "purchase_get",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/purchase_Purchase"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the token from setup or login."
      }
    },
    "schemas": {
...
      "product_Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "category_name": {
            "type": "string",
            "example": "category_name"
          },
          "name": {
            "type": "string",
            "example": "Coca-Cola 500ml"
          },
          "description": {
            "type": "string",
            "example": ""
          },
          "units_per_pack": {
            "type": "integer",
            "example": 24
          },
          "selling_price": {
            "type": "number",
            "example": 700
          },
          "average_cost_per_item": {
            "type": "number",
            "example": 0
          },
          "current_stock": {
            "type": "integer",
            "example": 0
          },
          "low_stock_level": {
            "type": "integer",
            "example": 20
          },
          "stock_status": {
            "type": "string",
            "example": "stock_status"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          }
        }
      },
      "product_CreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Coca-Cola 500ml",
            "minLength": 1
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "description": {
            "type": "string",
            "example": ""
          },
          "units_per_pack": {
            "type": "integer",
            "example": 24,
            "minimum": 1
          },
          "selling_price": {
            "type": "number",
            "example": 700,
            "minimum": 0
          },
          "low_stock_level": {
            "type": "integer",
            "example": 20,
            "minimum": 0
          }
        },
        "required": [
          "name",
          "category_id",
          "units_per_pack"
        ]
      },
      "product_UpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Coca-Cola 500ml",
            "minLength": 1
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "description": {
            "type": "string",
            "example": ""
          },
          "units_per_pack": {
            "type": "integer",
            "example": 24,
            "minimum": 1
          },
          "selling_price": {
            "type": "number",
            "example": 700,
            "minimum": 0
          },
          "low_stock_level": {
            "type": "integer",
            "example": 20,
            "minimum": 0
          }
        },
        "required": [
          "name",
          "category_id",
          "units_per_pack"
        ]
      },
...
      "purchase_CreateItemRequest": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "packs": {
            "type": "integer",
            "example": 10,
            "minimum": 1
          },
          "price_per_pack": {
            "type": "number",
            "example": 12000,
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": [
          "product_id",
          "packs",
          "price_per_pack"
        ]
      },
      "purchase_CreateRequest": {
        "type": "object",
        "properties": {
          "supplier_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "notes": {
            "type": "string",
            "example": ""
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/purchase_CreateItemRequest"
            },
            "nullable": false,
            "minItems": 1
          }
        },
        "required": [
          "supplier_id",
          "items"
        ]
      },
      "purchase_PurchaseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "product_name": {
            "type": "string",
            "example": "product_name"
          },
          "packs": {
            "type": "integer",
            "example": 10
          },
          "units_per_pack": {
            "type": "integer",
            "example": 24
          },
          "total_items": {
            "type": "integer",
            "example": 0
          },
          "price_per_pack": {
            "type": "number",
            "example": 12000
          },
          "price_per_item": {
            "type": "number",
            "example": 0
          },
          "total_cost": {
            "type": "number",
            "example": 0
          }
        }
      },
      "purchase_Purchase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "supplier_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "supplier_name": {
            "type": "string",
            "example": "supplier_name"
          },
          "purchase_date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          },
          "total_amount": {
            "type": "number",
            "example": 0
          },
          "notes": {
            "type": "string",
            "example": ""
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/purchase_PurchaseItem"
            },
            "nullable": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid request body"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed"
          }
        }
      },
...
    }
  }
}